Dinero es int, refactorizaciones varias
falta arreglar lo de los atributos publicos y que no dependamos de Posicion en Jugador

+++ src/modelo/interfaces/Adquirible.java
@@ -4,5 +4,5 @@ import modelo.Jugador;
 
 public interface Adquirible {
 	public abstract Jugador getDuenio();
-	public abstract double getPrecio();
+	public abstract int getPrecio();
 }
